chore(map): ratchet 250-NCL gates and densify map tests (Refs #4561)
Densify map package tests via support scenario harnesses, ratchet
repo.map_lib_file_size_headroom to 250 NCL and repo.map_test_file_size
to 250 physical lines, and align SPEC/repo gate docs (500 NCL hard cap,
barrel transitive export closure scan).
