+++ packages/colonizethis_map/test/support/init_game_map_view_fixtures.dart
@@ -15,6 +15,8 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 /// package `test/`; `colonizethis_test` gains no dependency on
 /// `colonizethis_map`. Refs #3746.
 
+export 'init_game_map_view_fixture_scenarios.dart';
+
 /// Builds a [TileMapResult] from a [grid], deriving width/height from the rows.
 ///
 /// All view-builder tests use rectangular grids, so width is taken from the
@@ -113,190 +115,5 @@ Game minimalGame({
     minorNations: minorNations,
     tribes: tribes,
     greatPowerColorOverride: greatPowerColorOverride,
-  );
-}
-
-/// Dual-region view-builder / visualizer scenario wiring.
-class DualRegionViewScenario {
-  const DualRegionViewScenario({
-    required this.game,
-    required this.tileMapByRegion,
-    required this.topologyByRegion,
-  });
-
-  final Game game;
-  final Map<String, TileMapResult> tileMapByRegion;
-  final Map<String, MapTopology> topologyByRegion;
-
-  InitGameMapViewData buildViewData({
-    int cellSize = 8,
-    int? seed,
-    String? configSummary,
-    Map<String, (int r, int g, int b)>? greatPowerColorOverride,
-    Map<String, TileVisibility>? visibilityByTile,
-    List<WarpLink>? warpLinks,
-    Map<String, int>? resourceExtractionUnitsByTile,
-    Map<String, int>? resourceExtractionEffectiveUnitsByTile,
-    Map<String, int>? resourceExtractionBlockedUnitsByTile,
-    Set<String>? civilianMarkerOwnerIds,
-    String? viewingFactionId,
-    Map<String, bool>? viewingTechUnlocked,
-  }) {
-    return buildInitGameMapViewData(
-      game: game,
-      tileMapByRegion: tileMapByRegion,
-      topologyByRegion: topologyByRegion,
-      cellSize: cellSize,
-      seed: seed,
-      configSummary: configSummary,
-      greatPowerColorOverride: greatPowerColorOverride,
-      visibilityByTile: visibilityByTile,
-      warpLinks: warpLinks,
-      resourceExtractionUnitsByTile: resourceExtractionUnitsByTile,
-      resourceExtractionEffectiveUnitsByTile:
-          resourceExtractionEffectiveUnitsByTile,
-      resourceExtractionBlockedUnitsByTile:
-          resourceExtractionBlockedUnitsByTile,
-      civilianMarkerOwnerIds: civilianMarkerOwnerIds,
-      viewingFactionId: viewingFactionId,
-      viewingTechUnlocked: viewingTechUnlocked,
-    );
-  }
-}
-
-/// Builds a [DualRegionViewScenario] from dual-region maps, topologies, and game.
-DualRegionViewScenario dualRegionViewScenario({
-  required Game game,
-  required TileMapResult oldWorldMap,
-  required TileMapResult newWorldMap,
-  required MapTopology oldWorldTopology,
-  required MapTopology newWorldTopology,
-}) {
-  return DualRegionViewScenario(
-    game: game,
-    tileMapByRegion: {'oldWorld': oldWorldMap, 'newWorld': newWorldMap},
-    topologyByRegion: {
-      'oldWorld': oldWorldTopology,
-      'newWorld': newWorldTopology,
-    },
-  );
-}
-
-/// Default single-tile new-world map for view-builder tests that focus on OW.
-TileMapResult defaultNewWorldMap() => mapTileGrid([
-  ['p1'],
-]);
-
-/// Default single-province new-world topology for view-builder tests.
-MapTopology defaultNewWorldTopology({List<String> provinceIds = const ['p1']}) {
-  return regionTopology(regionId: 'newWorld', provinceIds: provinceIds);
-}
-
-/// Shorthand [dualRegionViewScenario] with a default single-province new world.
-DualRegionViewScenario dualRegionScenario({
-  required Game game,
-  required List<List<String>> oldWorldGrid,
-  required MapTopology oldWorldTopology,
-  List<List<String>>? newWorldGrid,
-  MapTopology? newWorldTopology,
-  List<List<TerrainType?>>? oldWorldTerrainGrid,
-  List<List<Resource?>>? oldWorldResourceGrid,
-}) {
-  return dualRegionViewScenario(
-    game: game,
-    oldWorldMap: mapTileGrid(
-      oldWorldGrid,
-      terrainGrid: oldWorldTerrainGrid,
-      resourceGrid: oldWorldResourceGrid,
-    ),
-    newWorldMap: mapTileGrid(
-      newWorldGrid ??
-          const [
-            ['p1'],
-          ],
-    ),
-    oldWorldTopology: oldWorldTopology,
-    newWorldTopology: newWorldTopology ?? defaultNewWorldTopology(),
-  );
-}
-
-/// OW-focused scenario with a sea-only new-world placeholder grid.
-DualRegionViewScenario oldWorldFocusedScenario({
-  required Game game,
-  required List<List<String>> oldWorldGrid,
-  required MapTopology oldWorldTopology,
-  List<List<String>> newWorldGrid = const [
-    ['s1'],
-  ],
-  MapTopology? newWorldTopology,
-  List<List<TerrainType?>>? oldWorldTerrainGrid,
-}) {
-  return dualRegionViewScenario(
-    game: game,
-    oldWorldMap: mapTileGrid(oldWorldGrid, terrainGrid: oldWorldTerrainGrid),
-    newWorldMap: mapTileGrid(newWorldGrid),
-    oldWorldTopology: oldWorldTopology,
-    newWorldTopology:
-        newWorldTopology ?? singleProvinceAndSeaTopology('newWorld'),
-  );
-}
-
-/// Standard 2×2 province+sea grid in both regions (region/cell smoke tests).
-DualRegionViewScenario provinceSeaDualRegionScenario({required Game game}) {
-  final grid = [
-    ['p1', 's1'],
-    ['s1', 's1'],
-  ];
-  final topology = regionTopology(
-    regionId: 'oldWorld',
-    provinceIds: const ['p1'],
-    seaZoneIds: const ['s1'],
-    edges: const [TopologyEdge(id1: 'p1', id2: 's1')],
-  );
-  final nwTopology = regionTopology(
-    regionId: 'newWorld',
-    provinceIds: const ['p1'],
-    seaZoneIds: const ['s1'],
-    edges: const [TopologyEdge(id1: 'p1', id2: 's1')],
-  );
-  return dualRegionViewScenario(
-    game: game,
-    oldWorldMap: mapTileGrid(grid),
-    newWorldMap: mapTileGrid(grid),
-    oldWorldTopology: topology,
-    newWorldTopology: nwTopology,
-  );
-}
-
-/// Convenience wrapper for the common dual-region [buildInitGameMapViewData] call.
-InitGameMapViewData buildViewDataForScenario(
-  DualRegionViewScenario scenario, {
-  int cellSize = 8,
-  int? seed,
-  String? configSummary,
-  Map<String, (int r, int g, int b)>? greatPowerColorOverride,
-  Map<String, TileVisibility>? visibilityByTile,
-  List<WarpLink>? warpLinks,
-  Map<String, int>? resourceExtractionUnitsByTile,
-  Map<String, int>? resourceExtractionEffectiveUnitsByTile,
-  Map<String, int>? resourceExtractionBlockedUnitsByTile,
-  Set<String>? civilianMarkerOwnerIds,
-  String? viewingFactionId,
-  Map<String, bool>? viewingTechUnlocked,
-}) {
-  return scenario.buildViewData(
-    cellSize: cellSize,
-    seed: seed,
-    configSummary: configSummary,
-    greatPowerColorOverride: greatPowerColorOverride,
-    visibilityByTile: visibilityByTile,
-    warpLinks: warpLinks,
-    resourceExtractionUnitsByTile: resourceExtractionUnitsByTile,
-    resourceExtractionEffectiveUnitsByTile:
-        resourceExtractionEffectiveUnitsByTile,
-    resourceExtractionBlockedUnitsByTile: resourceExtractionBlockedUnitsByTile,
-    civilianMarkerOwnerIds: civilianMarkerOwnerIds,
-    viewingFactionId: viewingFactionId,
-    viewingTechUnlocked: viewingTechUnlocked,
   );
 }
