+++ packages/colonizethis_map/test/tile_map_gen_join_passes_test.dart
@@ -195,66 +195,11 @@ void main() {
   group('TerrainJitterPass', () {
     test(
       'jitter reassigns dominant edge cells toward supported neighbours',
-      () {
-        final params = genParams(
-          width: 5,
-          height: 5,
-          jitterMinProvinceSize: 4,
-          jitterHomogeneityThreshold: 0.5,
-          jitterMaxFraction: 1.0,
-          jitterProbability: 1.0,
-          jitterNeighborSupportThreshold: 1,
-        );
-        final pass = terrainJitterPassFor(params);
-        final grid = provinceOnlyGrid(5);
-        final terrain = plainsTerrainGrid(5);
-        terrain[2][2] = TerrainType.hills;
-        final resources = emptyResourceGrid(5);
-
-        final hillsBefore = countTerrainType(terrain, TerrainType.hills);
-        pass.jitterTerrainByProvince(
-          grid,
-          terrain,
-          resources,
-          'oldWorld',
-          Random(7),
-        );
-        final hillsAfter = countTerrainType(terrain, TerrainType.hills);
-        expect(
-          hillsAfter,
-          greaterThan(hillsBefore),
-          reason: 'plains edge cells adjacent to hills should flip to hills',
-        );
-      },
+      expectJitterReassignsDominantEdgeCells,
     );
 
     test('jitter leaves provinces below the min size untouched', () {
-      final params = genParams(
-        width: 5,
-        height: 5,
-        jitterMinProvinceSize: 1000,
-      );
-      final pass = terrainJitterPassFor(params);
-      final grid = provinceOnlyGrid(5);
-      final terrain = plainsTerrainGrid(5);
-      terrain[2][2] = TerrainType.hills;
-      final resources = emptyResourceGrid(5);
-      final before = [
-        for (final row in terrain) [...row],
-      ];
-      pass.run(
-        MapGenPassContext<TerrainJitterPassPayload>(
-          params: params,
-          payload: TerrainJitterPassPayload(
-            grid: grid,
-            terrainGrid: terrain,
-            resourceGrid: resources,
-            regionId: 'oldWorld',
-            rnd: Random(7),
-          ),
-        ),
-      );
-      expect(terrain, before);
+      expectJitterSkipsProvincesBelowMinSize();
     });
   });
 }
