+++ tool/check_map_lib_file_size_headroom.dart
@@ -1,4 +1,4 @@
-// Wave-6 headroom gate: colonizethis_map gen/view/render modules ≤300 NCL.
+// Wave-7 headroom gate: colonizethis_map gen/view/render modules ≤250 NCL.
 // SPEC/program/repo-lint.md (`repo.map_lib_file_size_headroom`, Refs #4112, #4297, #4371).
 //
 // Complements `repo.map_lib_file_size` (500 NCL hard cap) with a peer-aligned
@@ -13,7 +13,7 @@ import 'check_dart_file_non_comment_line_size.dart'
 import 'check_map_lib_file_size.dart' show mapLibFileSizeScanRoots;
 import 'ct_repo_lint_scan_contract.dart';
 
-const int mapLibFileSizeHeadroomCeiling = 300;
+const int mapLibFileSizeHeadroomCeiling = 250;
 
 /// Shrink-only allowlist during transition; remove entries as splits land.
 const List<String> mapLibFileSizeHeadroomGrandfathered = <String>[];
